pyproject.toml
[tool.poetry]
name = "curpage"
version = "0.1.0"
description = ""
authors = ["hanihoneykim <107341819+hanihoneykim@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
pillow = "^9.4.0"
django-admin-rangefilter = "^0.9.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
